ide_db

Module active_parameter

This module provides functionality for querying callable information about a token.

Structs

ActiveParameter

Functions

callable_for_arg_list

Returns a [hir::Callable] this token is a part of and its argument index of said callable.

callable_for_node

callable_for_token

Returns a [hir::Callable] this token is a part of and its argument index of said callable.

generic_def_for_node

into_comma 🔒