ide_db::defs

Enum Definition

pub enum Definition {

    Macro(Macro),
    Field(Field),
    TupleField(TupleField),
    Module(Module),
    Crate(Crate),
    Function(Function),
    Adt(Adt),
    Variant(Variant),
    Const(Const),
    Static(Static),
    Trait(Trait),
    TypeAlias(TypeAlias),
    SelfType(Impl),
    GenericParam(GenericParam),
    Local(Local),
    Label(Label),
    DeriveHelper(DeriveHelper),
    BuiltinType(BuiltinType),
    BuiltinLifetime(StaticLifetime),
    BuiltinAttr(BuiltinAttr),
    ToolModule(ToolModule),
    ExternCrateDecl(ExternCrateDecl),
    InlineAsmRegOrRegClass(()),
    InlineAsmOperand(InlineAsmOperand),

}

Variants

Macro(Macro)

Field(Field)

TupleField(TupleField)

Module(Module)

Crate(Crate)

Function(Function)

Adt(Adt)

Variant(Variant)

Const(Const)

Static(Static)

Trait(Trait)

TypeAlias(TypeAlias)

SelfType(Impl)

GenericParam(GenericParam)

Local(Local)

Label(Label)

DeriveHelper(DeriveHelper)

BuiltinType(BuiltinType)

BuiltinLifetime(StaticLifetime)

BuiltinAttr(BuiltinAttr)

ToolModule(ToolModule)

ExternCrateDecl(ExternCrateDecl)

InlineAsmRegOrRegClass(())

InlineAsmOperand(InlineAsmOperand)

Implementations

impl Definition

pub fn canonical_module_path( &self, db: &RootDatabase, ) -> Option<impl Iterator<Item = Module>>

pub fn krate(&self, db: &RootDatabase) -> Option<Crate>

pub fn module(&self, db: &RootDatabase) -> Option<Module>

Returns the module this definition resides in.

As such, for modules themselves this will return the parent module.

pub fn enclosing_definition(&self, db: &RootDatabase) -> Option<Definition>

pub fn visibility(&self, db: &RootDatabase) -> Option<Visibility>

pub fn name(&self, db: &RootDatabase) -> Option<Name>

pub fn docs( &self, db: &RootDatabase, famous_defs: Option<&FamousDefs<'_, '_>>, display_target: DisplayTarget, ) -> Option<Documentation>

pub fn docs_with_rangemap( &self, db: &RootDatabase, famous_defs: Option<&FamousDefs<'_, '_>>, display_target: DisplayTarget, ) -> Option<(Documentation, Option<DocsRangeMap>)>

pub fn label(&self, db: &RootDatabase, display_target: DisplayTarget) -> String

impl Definition

pub fn rename( &self, sema: &Semantics<'_, RootDatabase>, new_name: &str, rename_definition: RenameDefinition, ) -> Result<SourceChange>

pub fn range_for_rename( self, sema: &Semantics<'_, RootDatabase>, ) -> Option<FileRange>

Textual range of the identifier which will change when renaming this Definition. Note that builtin types can’t be renamed and extern crate names will report its range, though a rename will introduce an alias instead.

impl Definition

pub fn usages<'a>(self, sema: &'a Semantics<'_, RootDatabase>) -> FindUsages<'a>