ide_db::famous_defs

Struct FamousDefs

pub struct FamousDefs<'a, 'b>(pub &'a Semantics<'b, RootDatabase>, pub Crate);

Helps with finding well-know things inside the standard library. This is somewhat similar to the known paths infra inside hir, but it different; We want to make sure that IDE specific paths don’t become interesting inside the compiler itself as well.

Note that, by default, rust-analyzer tests do not include core or std libraries. If you are writing tests for functionality using FamousDefs, you’d want to include minicore (see test_utils::MiniCore) declaration at the start of your tests:

//- minicore: iterator, ord, derive

Tuple Fields
0: &'a Semantics<'b, RootDatabase>1: Crate
Implementations

impl FamousDefs<'_, '_>

pub fn std(&self) -> Option<Crate>

pub fn core(&self) -> Option<Crate>

pub fn alloc(&self) -> Option<Crate>

pub fn test(&self) -> Option<Crate>

pub fn proc_macro(&self) -> Option<Crate>

pub fn core_cmp_Ord(&self) -> Option<Trait>

pub fn core_convert_FromStr(&self) -> Option<Trait>

pub fn core_convert_From(&self) -> Option<Trait>

pub fn core_convert_Into(&self) -> Option<Trait>

pub fn core_convert_TryFrom(&self) -> Option<Trait>

pub fn core_convert_TryInto(&self) -> Option<Trait>

pub fn core_convert_Index(&self) -> Option<Trait>

pub fn core_option_Option(&self) -> Option<Enum>

pub fn core_result_Result(&self) -> Option<Enum>

pub fn core_default_Default(&self) -> Option<Trait>

pub fn core_iter_Iterator(&self) -> Option<Trait>

pub fn core_iter_IntoIterator(&self) -> Option<Trait>

pub fn core_iter(&self) -> Option<Module>

pub fn core_ops_Deref(&self) -> Option<Trait>

pub fn core_ops_DerefMut(&self) -> Option<Trait>

pub fn core_convert_AsRef(&self) -> Option<Trait>

pub fn core_convert_AsMut(&self) -> Option<Trait>

pub fn core_borrow_Borrow(&self) -> Option<Trait>

pub fn core_borrow_BorrowMut(&self) -> Option<Trait>

pub fn core_ops_ControlFlow(&self) -> Option<Enum>

pub fn core_ops_Drop(&self) -> Option<Trait>

pub fn core_marker_Copy(&self) -> Option<Trait>

pub fn core_marker_Sized(&self) -> Option<Trait>

pub fn core_future_Future(&self) -> Option<Trait>

pub fn core_macros_builtin_derive(&self) -> Option<Macro>

pub fn core_mem_drop(&self) -> Option<Function>

pub fn core_macros_todo(&self) -> Option<Macro>

pub fn core_macros_unimplemented(&self) -> Option<Macro>

pub fn core_fmt_Display(&self) -> Option<Trait>

pub fn alloc_string_ToString(&self) -> Option<Trait>

pub fn builtin_crates(&self) -> impl Iterator<Item = Crate>