ide_db::imports

Module import_assets

Look up accessible paths for items.

Structs

CompleteInFlyimport

ImportAssets

A struct to find imports in the project, given a certain name (or its part) and the context.

ImportPathConfig

LocatedImport

An import (not necessary the only one) that corresponds a certain given PathImportCandidate. (the structure is not entirely correct, since there can be situations requiring two imports, see FIXME below for the details)

PathImportCandidate

Path import for a given name, qualified or not.

TraitImportCandidate

A trait import needed for a given associated item access. For some::path::SomeStruct::ASSOC_, contains the type of some::path::SomeStruct and ASSOC_ as the item name.

Enums

ImportCandidate

A candidate for import, derived during various IDE activities:

NameToImport

A name that will be used during item lookups.

Functions

assoc_to_item 🔒

filter_candidates_by_after_path 🔒

get_mod_path 🔒

item_as_assoc 🔒

item_for_path_search

item_for_path_search_assoc 🔒

path_applicable_imports 🔒

path_import_candidate 🔒

trait_applicable_items 🔒

validate_resolvable 🔒

Validates and builds an import for resolved_qualifier if the unresolved_qualifier appended to it resolves and there is a validate candidate after that.