ide_db::imports

Module merge_imports

Handle syntactic aspects of merging UseTrees.

Modules

version_sort 🔒

Enums

MergeBehavior

What type of merges are allowed.

NormalizationStyle

Style to follow when normalizing a use tree.

Functions

common_prefix

Traverses both paths until they differ, returning the common prefix of both.

eq_attrs

eq_visibility

get_single_subtree 🔒

path_is_self 🔒

path_len 🔒

path_segment_cmp 🔒

recursive_merge 🔒

Recursively merges rhs to lhs

recursive_normalize 🔒

Recursively normalizes a use tree and its subtrees (if any).

remove_subtree_if_only_self 🔒

try_merge_imports

Merge rhs into lhs keeping both intact. Returned AST is mutable.

try_merge_trees

Merge rhs into lhs keeping both intact. Returned AST is mutable.

try_merge_trees_mut 🔒

try_normalize_import

Normalizes a use item by:

try_normalize_use_tree

Normalizes a use tree (see try_normalize_import doc).

try_normalize_use_tree_mut

use_tree_cmp 🔒

Orders use trees following rustfmt’s version sorting algorithm for ordering imports.

use_tree_cmp_bin_search 🔒

Use tree comparison func for binary searching for merging.

use_tree_cmp_by_tree_list_glob_or_alias 🔒

Orders for use trees with equal paths (see use_tree_cmp for details about use tree ordering).