ide_db::path_transform

Struct PathTransform

pub struct PathTransform<'a> { /* private fields */ }

PathTransform substitutes path in SyntaxNodes in bulk.

This is mostly useful for IDE code generation. If you paste some existing code into a new context (for example, to add method overrides to an impl block), you generally want to appropriately qualify the names, and sometimes you might want to substitute generic parameters as well:

mod x {
  pub struct A<V>;
  pub trait T<U> { fn foo(&self, _: U) -> A<U>; }
}

mod y {
  use x::T;

  impl T<()> for () {
     // If we invoke **Add Missing Members** here, we want to copy-paste `foo`.
     // But we want a slightly-modified version of it:
     fn foo(&self, _: ()) -> x::A<()> {}
  }
}

Implementations

impl<'a> PathTransform<'a>

pub fn trait_impl( target_scope: &'a SemanticsScope<'a>, source_scope: &'a SemanticsScope<'a>, trait_: Trait, impl_: Impl, ) -> PathTransform<'a>

pub fn function_call( target_scope: &'a SemanticsScope<'a>, source_scope: &'a SemanticsScope<'a>, function: Function, generic_arg_list: GenericArgList, ) -> PathTransform<'a>

pub fn impl_transformation( target_scope: &'a SemanticsScope<'a>, source_scope: &'a SemanticsScope<'a>, impl_: Impl, generic_arg_list: GenericArgList, ) -> PathTransform<'a>

pub fn adt_transformation( target_scope: &'a SemanticsScope<'a>, source_scope: &'a SemanticsScope<'a>, adt: Adt, generic_arg_list: GenericArgList, ) -> PathTransform<'a>

pub fn generic_transformation( target_scope: &'a SemanticsScope<'a>, source_scope: &'a SemanticsScope<'a>, ) -> PathTransform<'a>

pub fn apply(&self, syntax: &SyntaxNode) -> SyntaxNode

pub fn apply_all<'b>( &self, nodes: impl IntoIterator<Item = &'b SyntaxNode>, ) -> Vec<SyntaxNode>