ide_db
Module prime_caches

rust-analyzer is lazy and doesn’t compute anything unless asked. This sometimes is counter productive when, for example, the first goto definition request takes longer to compute. This module implements prepopulation of various caches, it’s not really advanced at the moment.

Structs

ParallelPrimeCachesProgress

We’re indexing many crates.

Functions

parallel_prime_caches