ide_db

Module search

Implementation of find-usages functionality.

It is based on the standard ide trick: first, we run a fast text search to get a super-set of matches. Then, we confirm each match using precise name resolution.

Structs

FileReference

FindUsages

ReferenceCategory

SearchScope

Generally, search_scope returns files that might contain references for the element. For pub(crate) things it’s a crate, for pub things it’s a crate and dependant crates. In some cases, the location of the references is known to within a TextRange, e.g. for things like local variables.

UsageSearchResult

Enums

FileReferenceNode

Functions

def_to_ty 🔒

is_name_ref_in_import 🔒

is_name_ref_in_test 🔒