ide_db

Module source_change

This modules defines type to represent changes to the source code, that flow from the server to the client.

It can be viewed as a dual for Change.

Structs

ChangeAnnotation

ChangeAnnotationId

An annotation ID associated with an indel, to describe changes.

SnippetBuilder

SnippetEdit

SourceChange

SourceChangeBuilder

TreeMutator

Enums

AnnotationSnippet

FileSystemEdit

Snippet