Module ide_db::syntax_helpers

Modules

format_string

Tools to work with format string literals for the format_args! family of macros.

format_string_exprs

Tools to work with expressions present in format string literals for the format_args! family of macros. Primarily meant for assists and completions.

insert_whitespace_into_node

Utilities for formatting macro expanded nodes until we get a proper formatter.

node_ext

Various helper functions to work with SyntaxNodes.

Structs

LexedStr