ide_db::syntax_helpers

Module suggest_name

This module contains functions to suggest names for expressions, functions and other items

Structs

NameGenerator

Generator for new names

Constants

SEQUENCE_TYPES 🔒

Generic types replaced by a plural of their first argument.

USELESS_METHODS 🔒

Useless methods that are stripped from expression

USELESS_METHOD_PREFIXES 🔒

Prefixes to strip from methods names

USELESS_NAMES 🔒

Identifier names that won’t be suggested, ever

USELESS_NAME_PREFIXES 🔒

USELESS_TRAITS 🔒

Trait names, that will be ignored when in impl Trait and dyn Trait

WRAPPER_TYPES 🔒

Generic types replaced by their first argument

Functions

from_call 🔒

from_field_name 🔒

from_func_call 🔒

from_method_call 🔒

from_param 🔒

from_type 🔒

is_useless_method 🔒

is_valid_name 🔒

name_of_type 🔒

normalize 🔒

sequence_name 🔒

trait_name 🔒

var_name_from_pat 🔒