ide_db::syntax_helpers::suggest_name
Struct NameGenerator

pub struct NameGenerator { /* private fields */ }

Generator for new names

The generator keeps track of existing names and suggests new names that do not conflict with existing names.

The generator will try to resolve conflicts by adding a numeric suffix to the name, e.g. a, a1, a2, …

Examples

let mut generator = NameGenerator::new();
assert_eq!(generator.suggest_name("a"), "a");
assert_eq!(generator.suggest_name("a"), "a1");

assert_eq!(generator.suggest_name("b2"), "b2");
assert_eq!(generator.suggest_name("b"), "b3");

Implementations

impl NameGenerator

pub fn new() -> Self

Create a new empty generator

pub fn new_with_names<'a>(existing_names: impl Iterator<Item = &'a str>) -> Self

Create a new generator with existing names. When suggesting a name, it will avoid conflicts with existing names.

pub fn new_from_scope_locals(scope: Option<SemanticsScope<'_>>) -> Self

pub fn suggest_name(&mut self, name: &str) -> SmolStr

Suggest a name without conflicts. If the name conflicts with existing names, it will try to resolve the conflict by adding a numeric suffix.

pub fn for_type( &mut self, ty: &Type, db: &RootDatabase, edition: Edition, ) -> Option<SmolStr>

Suggest a name for given type.

The function will strip references first, and suggest name from the inner type.

If ty is an ADT, it will suggest the name of the ADT.

If ty is wrapped in Box, Option or Result, it will suggest the name from the inner type.

If ty is a trait, it will suggest the name of the trait.

If ty is an impl Trait, it will suggest the name of the first trait.

If the suggested name conflicts with reserved keywords, it will return None.

pub fn for_impl_trait_as_generic(&mut self, ty: &ImplTraitType) -> SmolStr

Suggest name of impl trait type

Current implementation

In current implementation, the function tries to get the name from the first character of the name for the first type bound.

If the name conflicts with existing generic parameters, it will try to resolve the conflict with for_unique_generic_name.

pub fn for_variable( &mut self, expr: &Expr, sema: &Semantics<'_, RootDatabase>, ) -> SmolStr

Suggest name of variable for given expression

In current implementation, the function tries to get the name from the following sources:

if expr is an argument to function/method, use parameter name

if expr is a function/method call, use function name

expression type name if it exists (E.g. (), fn() -> () or ! do not have names)

fallback: var_name

It also applies heuristics to filter out less informative names

Currently it sticks to the first name found.