ide_db

Module traits

Functionality for obtaining data related to traits from the DB.

Functions

as_trait_assoc_def 🔒

If this is an trait (impl) assoc item, returns the assoc item of the corresponding trait definition.

assoc_item_of_trait 🔒

convert_to_def_in_trait 🔒

Converts associated trait impl items to their trait definition counterpart

get_missing_assoc_items

Given the impl block, returns the list of associated items (e.g. functions or types) that are missing in this impl block.

resolve_target_trait

Given the impl block, attempts to find the trait this impl corresponds to.