Module ide_db::defs

NameDefinition keeps information about the element we want to search references for. The element is represented by NameKind. It’s located inside some container and has a visibility, which defines a search scope. Note that the reference search is possible for not all of the classified items.

Enums

Definition

IdentClass

NameClass

On a first blush, a single ast::Name defines a single definition at some scope. That is, that, by just looking at the syntactical category, we can unambiguously define the semantic category.

NameRefClass

This is similar to NameClass, but works for [ast::NameRef] rather than for [ast::Name]. Similarly, what looks like a reference in syntax is a reference most of the time, but there are a couple of annoying exceptions.

OperatorClass