Module ide_db::helpers

Random assortment of ide helpers for high-level ide features that don’t fit in any other module.

Functions

get_definition

is_editable_crate

item_name

lint_eq_or_in_group

Checks if the given lint is equal or is contained by the other lint which may or may not be a group.

mod_path_to_ast

Converts the mod path struct into its ast representation.

pick_best_token

Picks the token with the highest rank returned by the passed in function.

pick_token

visit_file_defs

Iterates all ModuleDefs and Impl blocks of the given file.