Module ide_db::helpers

A module with ide helpers for high-level ide features.

Modules

import_assets
Look up accessible paths for items.

insert_use
Handle syntactic aspects of inserting a new use.

merge_imports
Handle syntactic aspects of merging UseTrees.

rust_doc
Rustdoc specific doc comment handling

Structs

FamousDefs
Helps with finding well-know things inside the standard library. This is somewhat similar to the known paths infra inside hir, but it different; We want to make sure that IDE specific paths don’t become interesting inside the compiler itself as well.

SnippetCap

Functions

item_name

mod_path_to_ast
Converts the mod path struct into its ast representation.

visit_file_defs
Iterates all ModuleDefs and Impl blocks of the given file.