Module ide_db::imports::import_assets

Look up accessible paths for items.

Structs

FirstSegmentUnresolved

A qualifier that has a first segment and it’s unresolved.

ImportAssets

A struct to find imports in the project, given a certain name (or its part) and the context.

LocatedImport

An import (not necessary the only one) that corresponds a certain given PathImportCandidate. (the structure is not entirely correct, since there can be situations requiring two imports, see FIXME below for the details)

PathImportCandidate

Path import for a given name, qualified or not.

TraitImportCandidate

A trait import needed for a given associated item access. For some::path::SomeStruct::ASSOC_, contains the type of some::path::SomeStruct and ASSOC_ as the item name.

Enums

ImportCandidate

A candidate for import, derived during various IDE activities:

NameToImport

A name that will be used during item lookups.

Functions

item_for_path_search