Module ide_db::imports::insert_use

Handle syntactic aspects of inserting a new use item.

Structs

InsertUseConfig

Enums

ImportGranularity

How imports should be grouped into use statements.

ImportScope

PrefixKind

Functions

insert_use

Insert an import path into the given file/node. A merge value of none indicates that no import merging is allowed to occur.

remove_path_if_in_use_stmt