ide_db

Module items_locator

This module has the functionality to search the project and its dependencies for a certain item, by its name and a few criteria. The main reason for this module to exist is the fact that project’s items and dependencies’ items are located in different caches, with different APIs.

Enums

AssocSearchMode

Three possible ways to search for the name in associated and/or other items.

Constants

DEFAULT_QUERY_SEARCH_LIMIT

A value to use, when uncertain which limit to pick.

Functions

items_with_name

Searches for importable items with the given name in the crate and its dependencies.

items_with_name_in_module

Searches for importable items with the given name in the crate and its dependencies.