Struct ide_db::source_change::SourceChangeBuilder

pub struct SourceChangeBuilder {
    pub edit: TextEditBuilder,
    pub file_id: FileId,
    pub source_change: SourceChange,
    pub trigger_signature_help: bool,
    pub mutated_tree: Option<TreeMutator>,
}

Fields
edit: TextEditBuilderfile_id: FileIdsource_change: SourceChangetrigger_signature_help: boolmutated_tree: Option<TreeMutator>

Maps the original, immutable SyntaxNode to a clone_for_update twin.

Implementations

impl SourceChangeBuilder

pub fn new(file_id: FileId) -> SourceChangeBuilder

pub fn edit_file(&mut self, file_id: FileId)

pub fn make_mut<N: AstNode>(&mut self, node: N) -> N

pub fn make_syntax_mut(&mut self, node: SyntaxNode) -> SyntaxNode

Returns a copy of the node, suitable for mutation.

Syntax trees in rust-analyzer are typically immutable, and mutating operations panic at runtime. However, it is possible to make a copy of the tree and mutate the copy freely. Mutation is based on interior mutability, and different nodes in the same tree see the same mutations.

The typical pattern for an assist is to find specific nodes in the read phase, and then get their mutable couterparts using make_mut in the mutable state.

pub fn delete(&mut self, range: TextRange)

Remove specified range of text.

pub fn insert(&mut self, offset: TextSize, text: impl Into<String>)

Append specified text at the given offset

pub fn insert_snippet(
&mut self,
_cap: SnippetCap,
offset: TextSize,
snippet: impl Into<String>
)

Append specified snippet at the given offset

pub fn replace(&mut self, range: TextRange, replace_with: impl Into<String>)

Replaces specified range of text with a given string.

pub fn replace_snippet(
&mut self,
_cap: SnippetCap,
range: TextRange,
snippet: impl Into<String>
)

Replaces specified range of text with a given snippet.

pub fn replace_ast<N: AstNode>(&mut self, old: N, new: N)

pub fn create_file(&mut self, dst: AnchoredPathBuf, content: impl Into<String>)

pub fn move_file(&mut self, src: FileId, dst: AnchoredPathBuf)

pub fn trigger_signature_help(&mut self)

pub fn finish(self) -> SourceChange