Trait ide_db::symbol_index::SymbolsDatabase

pub trait SymbolsDatabase: Database + HasQueryGroup<SymbolsDatabaseStorage> + HirDatabase + SourceDatabaseExt {
    fn file_symbols(&self, key0: FileId) -> Arc<SymbolIndex>;

    fn library_symbols(&self) -> Arc<FxHashMap<SourceRootId, SymbolIndex>>;

    fn local_roots(&self) -> Arc<FxHashSet<SourceRootId>>;

    fn set_local_roots(&mut self, value__: Arc<FxHashSet<SourceRootId>>);

    fn set_local_roots_with_durability(
        &mut self, 
        value__: Arc<FxHashSet<SourceRootId>>, 
        durability__: Durability
    );

    fn library_roots(&self) -> Arc<FxHashSet<SourceRootId>>;

    fn set_library_roots(&mut self, value__: Arc<FxHashSet<SourceRootId>>);

    fn set_library_roots_with_durability(
        &mut self, 
        value__: Arc<FxHashSet<SourceRootId>>, 
        durability__: Durability
    );
}

Required methods

fn file_symbols(&self, key0: FileId) -> Arc<SymbolIndex>

fn library_symbols(&self) -> Arc<FxHashMap<SourceRootId, SymbolIndex>>

fn local_roots(&self) -> Arc<FxHashSet<SourceRootId>>

The set of “local” (that is, from the current workspace) roots. Files in local roots are assumed to change frequently.

fn set_local_roots(&mut self, value__: Arc<FxHashSet<SourceRootId>>)

Set the value of the local_roots input.

See local_roots for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_local_roots_with_durability(
&mut self,
value__: Arc<FxHashSet<SourceRootId>>,
durability__: Durability
)

Set the value of the local_roots input and promise that its value will never change again.

See local_roots for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn library_roots(&self) -> Arc<FxHashSet<SourceRootId>>

The set of roots for crates.io libraries. Files in libraries are assumed to never change.

fn set_library_roots(&mut self, value__: Arc<FxHashSet<SourceRootId>>)

Set the value of the library_roots input.

See library_roots for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_library_roots_with_durability(
&mut self,
value__: Arc<FxHashSet<SourceRootId>>,
durability__: Durability
)

Set the value of the library_roots input and promise that its value will never change again.

See library_roots for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.