ide_db

Module syntax_helpers

Modules

format_string

Tools to work with format string literals for the format_args! family of macros.

format_string_exprs

Tools to work with expressions present in format string literals for the format_args! family of macros. Primarily meant for assists and completions.

node_ext

Various helper functions to work with SyntaxNodes.

suggest_name

This module contains functions to suggest names for expressions, functions and other items

tree_diff

Basic tree diffing functionality.

Structs

LexedStr

Functions

prettify_macro_expansion

Inserts whitespace and replaces $crate in macro expansions.