ide_db
Module text_edit

Representation of a TextEdit.

rust-analyzer never mutates text itself and only sends diffs to clients, so TextEdit is the ultimate representation of the work done by rust-analyzer.

Structs

Indel

InsertDelete – a single “atomic” change to text

TextEdit

TextEditBuilder

TextRange

A range in text, represented as a pair of TextSize.

TextSize

A measure of text length. Also, equivalently, an index into text.